[feature/master] 패키지 이름 변경 및 CompanyService 및 테스트 코드 수정.

+++ domain/src/main/kotlin/com/techinsights/domain/repository/company/CompanyRepositoryImpl.kt
@@ -45,9 +45,13 @@ class CompanyRepositoryImpl(
     return PageImpl(companyDtos, pageable, page.totalElements)
   }
 
-  override fun deleteById(id: String) {
+  override fun deleteById(id: String) : Boolean {
     val companyId = Tsid.decode(id)
-    companyJpaRepository.deleteById(companyId)
+    if (companyJpaRepository.existsById(companyId)) {
+      companyJpaRepository.deleteById(companyId)
+      return true
+    }
+    return false
   }
 
   override fun update(company: CompanyDto): CompanyDto {
